Fix type check for float value in _YamlDefaultValue.as_float
as_float func was checking for int type, probably due to copy-paste
error. This patch fixes the type check to float.

+++ moulin/yaml_wrapper.py
@@ -51,7 +51,7 @@ class _YamlDefaultValue:
     @property
     def as_float(self) -> float:
         "Get the floating point value"
-        if not isinstance(self._val, int):
+        if not isinstance(self._val, float):
             raise TypeError("Expected float value")
         return self._val
